newest changes to instruction object and added header file
