--- test.txt
@@ -1,67 +1,0 @@
-bbs     #0x80, 0x102a, 0x9a73
-lda     al, 0x106e
-cmp     al, #0x08
-bcc     0x9a73
-lda     al, 0x102f
-cmp     al, #0x64
-bcc     0x9a73
-ldx     0x10b4
-cpx     #0x000a
-bcc     0x9a73
-seb     #0x80, 0x102a
-bbs     #0x02, 0x105b, 0x9a49
-bbc     #0xff, 0x12c1, 0x9a73
-ldm     #0x00, 0x12c1
-bra     0x9a56
-lda     al, 0x102f
-cmp     al, 0x8aa3
-bcs     0x9f33
-ldx     #0x0000
-stx     0x1158
-stx     0x1164
-stx     0x11ca
-stx     0x1066
-stx     0x1068
-stx     0x106a
-lda     al, 0x102f
-cmp     al, 0x8aa3
-bcs     0x9f59
-ldx     #0x0000
-stx     0x1158
-stx     0x1164
-stx     0x11ca
-stx     0x1066
-stx     0x1068
-stx     0x106a
-ldx     #0x1130
-lda     al, 0x102f
-cmp     al, #0x8a
-bcc     0xa1b3
-bbs     #0x04, 0x105a, 0xa1b3
-bbs     #0x20, dp + 0x0f, 0xa196
-bbc     #0x40, 0x2000, 0xa18d
-jsr     0x99ff
-cmp     al, 0x8a97
-bcc     0xa18c
-seb     #0x08, 0x11ce
-bbs     #0x02, 0x1029, 0x9c3a
-ldx     #0x8af0
-lda     al, 0x1088
-cmp     al, dp + 0x01 + ix
-bcc     0x9c3a
-lda     al, 0x106e
-cmp     al, dp + 0x02 + ix
-bcc     0x9c3a
-lda     al, 0x1071
-cmp     al, dp + 0x03 + ix
-bcc     0x9c3a
-lda     al, 0x1282
-cmp     al, dp + 0x04 + ix
-bcc     0x9c3a
-ldx     0x1166
-bne     0x9c24
-stx     0x1166
-cpx     0x8af9
-bcc     0x9c40
-seb     #0x01, 0x1163
-bra     0x9c4b
